[ENH] created the front page of the application using the Aceternity UI and some custom components.
Theme change is applied in the application to toggle between light, dark and system theme.

Clerk Authentication is applied in development mode for the developers to test the sign-in, sign-up, manage account and Log out.

not-found js is added the src folder of the application to apply Not Found default page in the whole application.

Basic validation of user sign in and site home page url is implemented. if user tries to enter an invalid URL the application will go to the Not Found page.
If user tries to enter the url manually various validation for sign-in is implemented( refer middleware.ts file).